Put CONTEXT_SPAN_KEY in a separate class, to avoid exposing methods in ContextUtils.
After this commit, ContextUtils is split into two classes,
io.opencensus.trace.unsafe.ContextUtils and
io.opencensus.trace.CurrentSpanUtils.  ContextUtils has one public member,
CONTEXT_SPAN_KEY.  CurrentSpanUtils has internal methods for working with the
current Span, and it is package-private.
